// src/sub/types/const.ts
export enum ClashMode {
  DIRECT = "direct",
  GLOBAL = "global",
  RULE = "rule",
}

export enum DnsTag {
  FAKEIP = "dns:fakeip",
  LOCAL = "dns:local",
  PROXY = "dns:proxy",
  REJECT = "dns:reject",
}

export enum GeoIPTag {
  CN = "geoip:cn",
}

export enum GeoSiteTag {
  ADS = "geosite:ads",
  CN = "geosite:cn",
  PROXY = "geosite:proxy",
  PRIVATE = "geosite:private",
}

export enum OutboundTag {
  AI = "🤖 AI",
  AUTO = "🚀 Auto",
  BALANCED = "⭐ Balanced",
  DIRECT = "DIRECT",
  DNS = "out:dns",
  DOWNLOAD = "☁️ Download",
  EMBY = "🍟 Emby",
  GLOBAL = "🌐 Global",
  MEDIA = "📺 Media",
  PROXY = "PROXY",
  REJECT = "REJECT",
}

export enum RuleSetTag {
  ADS = "rule-set:ads",
  AI = "rule-set:ai",
  CN = "rule-set:cn",
  DOWNLOAD = "rule-set:download",
  EMBY = "rule-set:emby",
  MEDIA = "rule-set:media",
  PRIVATE = "rule-set:private",
  PROXY = "rule-set:proxy",
}